+++ client/src/context/authentication-context.tsx
@@ -1,25 +1,27 @@
 import React, { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-export const AuthContext = React.createContext({
+interface AuthContextModel {
+    isLoggedIn: boolean;
+    token?: string;
+    onLogout?: () => void;
+    onLogin?: (token: string) => void;
+}
+
+export const AuthContext = React.createContext<AuthContextModel>({
     isLoggedIn: false,
-    onLogout: () => {
-    },
-    onLogin: (token) => {
-    }
 });
 
-export const AuthContextProvider = (props) => {
+export const AuthContextProvider = (props: { children: React.ReactNode }) => {
     const [isLoggedIn, setIsLoggedIn] = useState(false);
     const [token, setToken] = useState('');
-
     const navigate = useNavigate();
 
     useEffect(() => {
         const storedUserLoggedInInformation = localStorage.getItem('isLoggedIn');
         if (storedUserLoggedInInformation === '1') {
             setIsLoggedIn(true);
-            setToken(localStorage.getItem('token'));
+            setToken(localStorage.getItem('token') || '');
         }
     }, []);
 
@@ -29,7 +31,7 @@ export const AuthContextProvider = (props) => {
         localStorage.removeItem('token');
     };
 
-    const loginHandler = (token) => {
+    const loginHandler = (token: string) => {
         setIsLoggedIn(true);
         setToken(token);
         localStorage.setItem('isLoggedIn', '1');
@@ -38,16 +40,16 @@ export const AuthContextProvider = (props) => {
     };
 
     return (
-        <AuthContext.Provider
-            value={{
-                isLoggedIn,
-                token,
-                onLogout: logoutHandler,
-                onLogin: loginHandler
-            }}
-        >
-            {props.children}
-        </AuthContext.Provider>
+      <AuthContext.Provider
+        value={{
+            isLoggedIn,
+            token,
+            onLogout: logoutHandler,
+            onLogin: loginHandler
+        }}
+      >
+          {props.children}
+      </AuthContext.Provider>
     );
 };
