+++ client/src/global-settings/colors.ts
