--- client/src/pages/ProfilePage.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const ProfilePage = () => {
-    return <div>Hi there</div>;
-};
-
-export default ProfilePage;
