+++ client/src/pages/ProfilePage.tsx
@@ -1,0 +1,7 @@
+import React, { FC } from 'react';
+
+const ProfilePage: FC = () => {
+    return <div>Hi there</div>;
+};
+
+export default ProfilePage;
